Fixed ESLint issue
