--- src/components/modal/Header/header.module.scss
@@ -1,10 +1,0 @@
-.container {
-  color: khaki;
-}
-
-.modalheader {
-  text-align: center;
-  padding-top: 0.625rem;
-  font-family: 'Arial';
-  color: #1d1283;
-}
