--- src/components/modal/Header/index.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import classNames from '@components/modal/Header/header.module.scss';
-
-const Header = () => (
-  <h1 className={classNames.modalheader}>Send Your Interest</h1>
-);
-
-export default Header;
